// packages/symbol/src/symbol/publicSymbol.interface.ts
export interface IPublicSymbol {
    (this: any, description?: string): any
    asyncIterator: any
    hasInstance: any
    isConcatSpreadable: any
    iterator: any
    match: any
    matchAll: any
    replace: any
    search: any
    species: any
    split: any
    toPrimitive: any
    toStringTag: any
    unscopables: any
}